tools/codis2pika/test/requirements.txt
jury-test==0.0.3
redis==4.3.3
toml==0.10.2
